test(canvas): stabilize retirement CI coverage
