+++ extensions/canvas/src/host/a2ui.test.ts
@@ -1,6 +1,29 @@
-import { describe, expect, it } from "vitest";
+import fs from "node:fs/promises";
+import path from "node:path";
+import { resolvePreferredOpenClawTmpDir } from "openclaw/plugin-sdk/temp-path";
+import { afterAll, beforeAll, describe, expect, it } from "vitest";
+import { handleA2uiHttpRequestWithRootResolver } from "./a2ui-route.js";
 import { A2UI_PATH } from "./a2ui-shared.js";
-import { handleA2uiHttpRequest } from "./a2ui.js";
+
+let fixtureRoot = "";
+let fixtureRootReal = "";
+
+beforeAll(async () => {
+  fixtureRoot = await fs.mkdtemp(
+    path.join(resolvePreferredOpenClawTmpDir(), "openclaw-canvas-a2ui-route-"),
+  );
+  const assetRoot = path.join(fixtureRoot, "canvas-host", "a2ui");
+  await fs.mkdir(assetRoot, { recursive: true });
+  await Promise.all([
+    fs.writeFile(path.join(assetRoot, "a2ui.bundle.js"), "window.v08 = true;", "utf8"),
+    fs.writeFile(path.join(assetRoot, "a2ui-v0.9.bundle.js"), "window.v09 = true;", "utf8"),
+  ]);
+  fixtureRootReal = await fs.realpath(assetRoot);
+});
+
+afterAll(async () => {
+  await fs.rm(fixtureRoot, { recursive: true, force: true });
+});
 
 type CapturedResponse = {
   handled: boolean;
@@ -9,7 +32,11 @@ type CapturedResponse = {
   body: Buffer;
 };
 
-async function capture(url: string, method = "GET"): Promise<CapturedResponse> {
+async function capture(
+  url: string,
+  method = "GET",
+  resolveRootReal: () => Promise<string | null> = async () => fixtureRootReal,
+): Promise<CapturedResponse> {
   const response: CapturedResponse = {
     handled: false,
     status: 200,
@@ -26,7 +53,11 @@ async function capture(url: string, method = "GET"): Promise<CapturedResponse> {
       response.body = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk ?? "");
     },
   };
-  response.handled = await handleA2uiHttpRequest({ method, url }, res);
+  response.handled = await handleA2uiHttpRequestWithRootResolver(
+    { method, url },
+    res,
+    resolveRootReal,
+  );
   response.status = res.statusCode;
   return response;
 }
@@ -57,5 +88,12 @@ describe("Canvas A2UI renderer asset route", () => {
       expect(response.body.toString("utf8")).toBe("not found");
     }
     expect((await capture(A2UI_PATH, "POST")).status).toBe(405);
+  });
+
+  it("returns unavailable when renderer assets cannot be resolved", async () => {
+    const response = await capture(`${A2UI_PATH}/a2ui.bundle.js`, "GET", async () => null);
+
+    expect(response.status).toBe(503);
+    expect(response.body.toString("utf8")).toBe("A2UI assets not found");
   });
 });
